+++ testing/test_session.py
@@ -1,5 +1,5 @@
 import pytest
-from _pytest.main import ExitCode
+from _pytest.config import ExitCode
 
 
 class SessionTests:
