+++ testing/test_terminal.py
@@ -12,7 +12,7 @@ import pluggy
 import py
 
 import pytest
-from _pytest.main import ExitCode
+from _pytest.config import ExitCode
 from _pytest.pytester import Testdir
 from _pytest.reports import BaseReport
 from _pytest.terminal import _folded_skips
